first statemachine attempt

+++ blaze-lite/core/lib/sdCard/sdcard.h
@@ -7,7 +7,7 @@
 class sdCard {
     public:
         //Constructors
-        sdCard(const int csPin);
+        sdCard(const uint8_t csPin);
         //Destructor
         ~sdCard(); 
         //setup function
@@ -20,6 +20,7 @@ class sdCard {
         ssize_t writeData(const size_t bytes, const char* data);
         ssize_t readData(const size_t bytes, char* buffer);
         //log read/write methods
+        //new comment
         ssize_t writeLog(const char* logEntry, const size_t length);
         ssize_t readLog(char* buffer, const size_t maxLength);
